# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pannaga API",
    "description": "Hyperlocal Monsoon Onset & Break Prediction System for India",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/advisories": {
      "get": {
        "tags": [
          "Advisories"
        ],
        "summary": "List advisories for the current farmer's panchayat",
        "description": "Returns APPROVED advisories for the farmer's panchayat (most recent first).\nIncludes a persistent `simulated_banner` flag when data_source=SIMULATED.",
        "operationId": "list_advisories_api_v1_advisories_get",
        "parameters": [
          {
            "name": "lead_week",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lead Week"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/advisories/{advisory_id}": {
      "get": {
        "tags": [
          "Advisories"
        ],
        "summary": "Advisory detail",
        "operationId": "get_advisory_api_v1_advisories__advisory_id__get",
        "parameters": [
          {
            "name": "advisory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Advisory Id"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/advisories/{advisory_id}/translate/{lang}": {
      "get": {
        "tags": [
          "Advisories"
        ],
        "summary": "On-demand translation",
        "operationId": "translate_advisory_endpoint_api_v1_advisories__advisory_id__translate__lang__get",
        "parameters": [
          {
            "name": "advisory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Advisory Id"
            }
          },
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/officer/advisories/{advisory_id}/translate/{lang}": {
      "get": {
        "tags": [
          "Officer"
        ],
        "summary": "Translation preview",
        "operationId": "translation_preview_api_v1_officer_advisories__advisory_id__translate__lang__get",
        "parameters": [
          {
            "name": "advisory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Advisory Id"
            }
          },
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}